+++ hps/app/pension/components/Slider.tsx
